prelude_msg_write now only return success or error, and handle non blocking internally. Update function prototype and documentation.
git-svn-id: file:///home/yoann/SVN/trunk/libprelude@4560 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/prelude-msg.h
@@ -56,7 +56,7 @@ prelude_msg_t *prelude_msg_new(size_t msgcount, size_t msglen, uint8_t tag, uint
 
 void prelude_msg_set(prelude_msg_t *msg, uint8_t tag, uint32_t len, const void *data);
 
-ssize_t prelude_msg_write(prelude_msg_t *msg, prelude_io_t *dst);
+int prelude_msg_write(prelude_msg_t *msg, prelude_io_t *dst);
